Changing the way that we are making the website
Changing from using jekyll pages to sphinx compiled
documentation pages. I believe that this will make it
much more visually appealing and the code will look a
lot neater on the page. Also, makes it a lot easier
to compile the pages and see what the site will look
like.

Currently reproduced the FD method page and some of
the contents on the main page. Will use a mixture of
Markdown and reStructuredTxt files. All code will be
either in a jupyter notebook to be converted into
Markdown with nbsphinx or straight rst.